Wallpaper created when ALAP
Wallpaper is now created as late as possible. When user taps "wallpaper"
button on UI, the wallpaper will be created. This takes some time and
the phone will "freeze" for a while. Later I'll try to implement a
loading screen of some sort. Also a toast appears when the wallpaper is
changed.